move decompiler test data into a separate file
